+++ Tests/EscriboCoreTests/FountainWriterTitlePageTests.swift
@@ -684,4 +684,3 @@ struct FountainWriterIdempotenceTests {
     }
   }
 }
-